+++ pkg/utils/test_utils.go
@@ -3,8 +3,6 @@ package utils
 import (
 	"github.com/onsi/gomega/format"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/client-go/tools/record"
 )
 
@@ -36,39 +34,4 @@ func (matcher AlreadyExistMatcher) FailureMessage(actual interface{}) (message s
 // NegatedFailureMessage builds an error message.
 func (matcher AlreadyExistMatcher) NegatedFailureMessage(actual interface{}) (message string) {
 	return format.Message(actual, "not to be already exist")
-}
-
-func upsertOwnerRef(ref metav1.OwnerReference, object metav1.Object) {
-	owners := object.GetOwnerReferences()
-	if idx := indexOwnerRef(owners, ref); idx == -1 {
-		owners = append(owners, ref)
-	} else {
-		owners[idx] = ref
-	}
-	object.SetOwnerReferences(owners)
-}
-
-// indexOwnerRef returns the index of the owner reference in the slice if found, or -1.
-func indexOwnerRef(ownerReferences []metav1.OwnerReference, ref metav1.OwnerReference) int {
-	for index, r := range ownerReferences {
-		if referSameObject(r, ref) {
-			return index
-		}
-	}
-	return -1
-}
-
-// Returns true if a and b point to the same object.
-func referSameObject(a, b metav1.OwnerReference) bool {
-	aGV, err := schema.ParseGroupVersion(a.APIVersion)
-	if err != nil {
-		return false
-	}
-
-	bGV, err := schema.ParseGroupVersion(b.APIVersion)
-	if err != nil {
-		return false
-	}
-
-	return aGV.Group == bGV.Group && a.Kind == b.Kind && a.Name == b.Name
 }
